requirements.txt
uvicorn
fastapi
mysql-connector-python
fastapi_users
fastapi_users[password]
pyjwt[crypto]
jose
python-jose[cryptography]
